Allow App Highlight without turning on Experimental Feature Flag https://bugs.webkit.org/show_bug.cgi?id=219668
Reviewed by Wenson Hsieh.

The same painting infrastructure is used to paint app highlights as well as JavaScript highlights, but the
both were gated on the experimental feature flag for the JavaScript highlights when painting text. The app
highlights should not have such a barrier. Conversely, painting the JavaScript highlights was not gated when
painting images. That should be behind the feature flag check.

* rendering/InlineTextBox.cpp:
(WebCore::InlineTextBox::collectMarkedTextsForHighlights const):
* rendering/RenderReplaced.cpp:
(WebCore::RenderReplaced::calculateHighlightColor const):

Identifier: 232256@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@270585 268f45cc-cd09-0410-ab3c-d52691b4dbfc