[GPU Process]: Recording an in-process ImageBuffer drawing has to convert it to a NativeImage first https://bugs.webkit.org/show_bug.cgi?id=219705
Reviewed by Wenson Hsieh.

Source/WebCore:

We will allow GraphicsContext::drawImageBuffer() to proceed with the
painting code path if the ImageBuffer is not a RemoteImageBuffer. In this
case ImageBuffer::draw() extracts a NativeImage from the ImageBuffer and
calls GraphicsContext::drawNativeImage() which will send the NativeImage
to GPUP.

* platform/graphics/GraphicsContext.cpp:
(WebCore::GraphicsContext::drawImageBuffer):
* platform/graphics/GraphicsContextImpl.h:
(WebCore::GraphicsContextImpl::canDrawImageBuffer const):
* platform/graphics/displaylists/DisplayListRecorder.cpp:
(WebCore::DisplayList::Recorder::canDrawImageBuffer const):
* platform/graphics/displaylists/DisplayListRecorder.h:
(WebCore::DisplayList::Recorder::Delegate::isCachedImageBuffer const):

Source/WebKit:

Override the virtual method isCachedImageBuffer() which in this case
answers the question: is this a remote ImageBuffer or not?

* WebProcess/GPU/graphics/RemoteImageBufferProxy.h:


Identifier: 232272@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@270605 268f45cc-cd09-0410-ab3c-d52691b4dbfc